re-organized to existing code structure
