+++ tests/index.html
@@ -68,7 +68,7 @@
     <script src="spec/test-variables-api-service.js"></script>
     <script src="spec/test-cookie-store.js"></script>
 
-    <script src="spec/test-channel.js"></script>
+    <script src="spec/test-channel-service.js"></script>
 
 
     <script>
