+++ tests/spec/test-channel-service.js
